hopefully fix column offset calculation #124
